+++ packages/populationCharacteristics/src/components/EditSettings/tests/utils.test.tsx
@@ -1,0 +1,7 @@
+import { store } from '@onaio/redux-reducer-registry';
+import { editSetting } from '../utils';
+import { setting1 } from './fixtures';
+
+/*describe('src/components/EditSettings/utils/onEditSuccess', () => {
+    it('updates store correctly for inherited setting', () => {});
+});*/
